+++ R/datasets.R
@@ -37,7 +37,6 @@
 #' \insertRef{REVE:TRAI:01}{mlogit}
 #' @keywords datasets
 #' @importFrom Rdpack reprompt
-
 NULL
 
 
@@ -234,13 +233,18 @@ NULL
 #' ModeCanada <- subset(ModeCanada, noalt == 4)
 #' ModeCanada <- subset(ModeCanada, alt != "bus")
 #' ModeCanada$alt <- ModeCanada$alt[drop = TRUE]
-#' KoppWen00 <- mlogit.data(ModeCanada, shape='long', chid.var = 'case',
-#'                          alt.var = 'alt', choice = 'choice',
-#'                          drop.index = TRUE)
+#' KoppWen00 <- dfidx(ModeCanada)
+#'
 #' pcl <- mlogit(choice ~ freq + cost + ivt + ovt, KoppWen00, reflevel = 'car',
 #'               nests = 'pcl', constPar = c('iv:train.air'))
 #' 
 NULL
+
+
+## #' KoppWen00 <- mlogit.data(ModeCanada, shape='long', chid.var = 'case',
+## #'                          alt.var = 'alt', choice = 'choice',
+## #'                          drop.index = TRUE)
+
 
 #' Technologies to reduce NOx emissions
 #'